requirements.txt
ahttp==1.1.1
amqp==5.0.5
aiofiles==0.7.0
asgiref==3.3.1
bcrypt==3.2.0
fastapi==0.63.0
Jinja2==2.11.3
loguru==0.5.3
pypinyin==0.43.0
passlib==1.7.4
pydantic==1.7.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
redis==3.5.3
requests==2.5.0
rsa==4.7
SQLAlchemy==1.3.22
starlette==0.13.6
CMRESHandler==1.0.0
email-validator==1.1.3
jwt
PyJWT
fastapi
uvicorn
gunicorn

